fixed components folder
